State's gameVariables are now doubles, POSITION_X/_Y/_Z gameVariables added (#127), objects' positions added to labels
Former-commit-id: 9fe51f667b0158b675e53b9a54f12deb85209488

+++ src/lib/ViZDoomController.h
@@ -234,7 +234,7 @@ namespace vizdoom{
         /* GameState getters */
         /*------------------------------------------------------------------------------------------------------------*/
 
-        int getGameVariable(GameVariable var);
+        double getGameVariable(GameVariable var);
 
         unsigned int getGameTic();
         bool isMultiplayerGame();
